Move runtime memory identities into core
