+++ crates/trust-runtime-core/src/memory.rs
@@ -1,0 +1,57 @@
+//! Portable runtime memory identity types.
+
+/// Memory location identifier.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum MemoryLocation {
+    /// Global variable area.
+    Global,
+    /// Local variable area for a specific call frame.
+    Local(FrameId),
+    /// FB/Class instance storage.
+    Instance(InstanceId),
+    /// I/O area (direct addresses).
+    Io(IoArea),
+    /// Retain area (persistent across warm restart).
+    Retain,
+}
+
+/// I/O area identifiers per IEC 61131-3.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum IoArea {
+    /// Input area (%I).
+    Input,
+    /// Output area (%Q).
+    Output,
+    /// Memory area (%M).
+    Memory,
+}
+
+/// Frame identifier for call stack.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub struct FrameId(pub u32);
+
+/// Instance identifier for FB/Class instances.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub struct InstanceId(pub u32);
+
+#[cfg(test)]
+mod tests {
+    use super::{FrameId, InstanceId, IoArea, MemoryLocation};
+
+    #[test]
+    fn memory_identity_values_preserve_equality_and_hash_shape() {
+        assert_eq!(MemoryLocation::Global, MemoryLocation::Global);
+        assert_eq!(
+            MemoryLocation::Local(FrameId(7)),
+            MemoryLocation::Local(FrameId(7))
+        );
+        assert_ne!(
+            MemoryLocation::Instance(InstanceId(1)),
+            MemoryLocation::Instance(InstanceId(2))
+        );
+        assert_eq!(
+            MemoryLocation::Io(IoArea::Input),
+            MemoryLocation::Io(IoArea::Input)
+        );
+    }
+}
