+++ docs/internal/testing/checklists/architecture-improvements.md
@@ -40,6 +40,7 @@ Issue #51 init benchmark baseline evidence (2026-04-28):
 - [x] `ARCH-RTCORE-04` Close runtime-core split Phase 1 behavior locks and Phase 2 pre-extraction doctor rules before production code movement: VM/cycle/scheduler/retain/watchdog/initializer behavior locks pass, `FULLMAP-CHECK-05` covers the full runtime-core forbidden dependency/import set, `FULLMAP-CHECK-10` now reports function-size advisories and fails oversized `trust-runtime-core` functions, and `cargo run -p xtask -- architecture-doctor --full-map` passes at `62156e7e9`.
 - [x] `ARCH-RTCORE-05` Add the initial `trust-runtime-core` scaffold as a workspace member with default `std`, `no_std`-shaped crate attributes, ownership/exclusion docs, one `scaffold` module, one pre-move unit test, and an explicit `trust-runtime::runtime_core` compatibility re-export; `FULLMAP-CHECK-05` passes with the scaffold present and no forbidden dependencies/imports.
 - [x] `ARCH-RTCORE-06` Start Phase 4 value-model movement with the standalone date/time primitives in `trust-runtime-core`, preserving the existing `trust-runtime::value` re-export path and validating the slice with `cargo test -p trust-runtime-core -- --nocapture`, `cargo check -p trust-runtime --lib`, and `cargo check -p trust-runtime-core --no-default-features`.
+- [x] `ARCH-RTCORE-07` Move portable memory identity types (`MemoryLocation`, `IoArea`, `FrameId`, `InstanceId`) into `trust-runtime-core` while preserving `trust-runtime::memory` re-exports; validate with runtime-core tests, focused runtime memory tests, no-default core check, clippy for touched crates, and the full-map doctor dependency fence.
 - [x] `ARCH-FULL-01` Add `full-architecture-refactor-program-checklist.md` as the umbrella board that states the runtime-core split is one branch of the full SOLID/KISS and zero-silent-bug program, not the whole program.
 - [x] `ARCH-FULL-02` Add `architecture-doctor-full-map-execution-checklist.md` so `architecture-doctor --full-map` becomes a concrete prerequisite for enforcing workspace, runtime-core, CLI, host-surface, KISS, API, and diagram rules.
 - [x] `ARCH-FULL-03` Add `hir-mutation-hardening-execution-checklist.md` for the F2 silent-bug gap: `symbol_import`, `type_check::const_eval`, and aggregate initializer validation mutation survivors.
